--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.3.0] - 2020-07-29
+## [1.3.1] - 2020-07-30
+
+### Changed
+- Tagref is now faster because it now compiles the regular expressions once, rather than once per file.
+
+## [1.3.0] - 2020-07-30
 
 ### Added
 - Added support for `--ref-prefix` and `--tag-prefix`.