--- a/Cargo.toml
+++ b/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "tagref"
-version = "1.3.0"
+version = "1.3.1"
 authors = ["Stephan Boyer <stephan@stephanboyer.com>"]
 description = "Tagref helps you refer to other locations in your codebase."
 license = "MIT"