--- a/src/label.rs
+++ b/src/label.rs
@@ -1,4 +1,4 @@
-use regex::{escape, Regex};
+use regex::Regex;
 use std::{
     fmt,
     io::BufRead,
@@ -38,27 +38,19 @@ impl fmt::Display for Label {
 
 // This function returns all the labels in a file for a given type.
 pub fn parse<R: BufRead>(
-    tag_prefix: &str,
-    ref_prefix: &str,
+    tag_regex: &Regex,
+    ref_regex: &Regex,
     label_type: Type,
     path: &Path,
     reader: R,
 ) -> Vec<Label> {
-    // The `unwrap`s below are safe because we manually inspected the regexes.
     let regex = match label_type {
-        Type::Tag => Regex::new(&format!(
-            "(?i)\\[\\s*{}\\s*:\\s*([^\\]\\s]*)\\s*\\]",
-            escape(tag_prefix)
-        ))
-        .unwrap(),
-        Type::Ref => Regex::new(&format!(
-            "(?i)\\[\\s*{}\\s*:\\s*([^\\]\\s]*)\\s*\\]",
-            escape(ref_prefix)
-        ))
-        .unwrap(),
+        Type::Tag => tag_regex,
+        Type::Ref => ref_regex,
     };
 
     let mut labels: Vec<Label> = Vec::new();
+
     for (line_number, line_result) in reader.lines().enumerate() {
         if let Ok(line) = line_result {
             for captures in regex.captures_iter(&line) {
@@ -80,14 +72,21 @@ pub fn parse<R: BufRead>(
 #[cfg(test)]
 mod tests {
     use crate::label::{parse, Type};
+    use regex::Regex;
     use std::path::Path;
+
+    const TAG_REGEX: &str = "(?i)\\[\\s*tag\\s*:\\s*([^\\]\\s]*)\\s*\\]";
+    const REF_REGEX: &str = "(?i)\\[\\s*ref\\s*:\\s*([^\\]\\s]*)\\s*\\]";
 
     #[test]
     fn parse_empty() {
         let path = Path::new("file.rs").to_owned();
         let contents = b"" as &[u8];
 
-        let tags = parse("tag", "ref", Type::Tag, &path, contents);
+        let tag_regex: Regex = Regex::new(TAG_REGEX).unwrap();
+        let ref_regex: Regex = Regex::new(REF_REGEX).unwrap();
+
+        let tags = parse(&tag_regex, &ref_regex, Type::Tag, &path, contents);
 
         assert!(tags.is_empty());
     }
@@ -101,7 +100,10 @@ mod tests {
         .trim()
         .as_bytes();
 
-        let tags = parse("tag", "ref", Type::Tag, &path, contents);
+        let tag_regex: Regex = Regex::new(TAG_REGEX).unwrap();
+        let ref_regex: Regex = Regex::new(REF_REGEX).unwrap();
+
+        let tags = parse(&tag_regex, &ref_regex, Type::Tag, &path, contents);
 
         assert_eq!(tags.len(), 1);
         assert_eq!(tags[0].label_type, Type::Tag);
@@ -119,7 +121,10 @@ mod tests {
         .trim()
         .as_bytes();
 
-        let refs = parse("tag", "ref", Type::Ref, &path, contents);
+        let tag_regex: Regex = Regex::new(TAG_REGEX).unwrap();
+        let ref_regex: Regex = Regex::new(REF_REGEX).unwrap();
+
+        let refs = parse(&tag_regex, &ref_regex, Type::Ref, &path, contents);
 
         assert_eq!(refs.len(), 1);
         assert_eq!(refs[0].label_type, Type::Ref);
@@ -137,7 +142,10 @@ mod tests {
         .trim()
         .as_bytes();
 
-        let tags = parse("tag", "ref", Type::Tag, &path, contents);
+        let tag_regex: Regex = Regex::new(TAG_REGEX).unwrap();
+        let ref_regex: Regex = Regex::new(REF_REGEX).unwrap();
+
+        let tags = parse(&tag_regex, &ref_regex, Type::Tag, &path, contents);
 
         assert_eq!(tags.len(), 1);
         assert_eq!(tags[0].label_type, Type::Tag);
@@ -156,7 +164,10 @@ mod tests {
         .trim()
         .as_bytes();
 
-        let tags = parse("tag", "ref", Type::Tag, &path, contents);
+        let tag_regex: Regex = Regex::new(TAG_REGEX).unwrap();
+        let ref_regex: Regex = Regex::new(REF_REGEX).unwrap();
+
+        let tags = parse(&tag_regex, &ref_regex, Type::Tag, &path, contents);
 
         assert_eq!(tags.len(), 2);
         assert_eq!(tags[0].label_type, Type::Tag);
@@ -179,7 +190,10 @@ mod tests {
         .trim()
         .as_bytes();
 
-        let tags = parse("tag", "ref", Type::Tag, &path, contents);
+        let tag_regex: Regex = Regex::new(TAG_REGEX).unwrap();
+        let ref_regex: Regex = Regex::new(REF_REGEX).unwrap();
+
+        let tags = parse(&tag_regex, &ref_regex, Type::Tag, &path, contents);
 
         assert_eq!(tags.len(), 2);
         assert_eq!(tags[0].label_type, Type::Tag);
@@ -202,7 +216,10 @@ mod tests {
         .trim()
         .as_bytes();
 
-        let tags = parse("tag", "ref", Type::Tag, &path, contents);
+        let tag_regex: Regex = Regex::new(TAG_REGEX).unwrap();
+        let ref_regex: Regex = Regex::new(REF_REGEX).unwrap();
+
+        let tags = parse(&tag_regex, &ref_regex, Type::Tag, &path, contents);
 
         assert_eq!(tags.len(), 2);
         assert_eq!(tags[0].label_type, Type::Tag);