Precompile regexes to improve performance.

Before:

```sh
tagref  8.17s user 3.61s system 96% cpu 12.192 total
```

After:

```sh
tagref  3.43s user 3.49s system 68% cpu 10.104 total
```

These are sample runs, but they are adequately representative of the average of a few runs.

**Status:** Ready

**Fixes:** N/A